added styles to the Edit Booking page
