+++ src/auth/components/signUpForm/SignUpButton.tsx
@@ -1,13 +1,13 @@
 import Button from '@mui/material/Button';
 import React from 'react';
 import { Link } from 'react-router-dom';
-import { signUpForm } from 'src/router/rootConstants';
+import { signUpFormPath } from 'src/router/rootConstants';
 
 export const SignUpButton = () => {
   return (
     <div>
       <Button variant="contained" type="submit" size={'small'}>
-        <Link to={signUpForm}>Зарегистрироваться</Link>
+        <Link to={signUpFormPath}>Зарегистрироваться</Link>
       </Button>
     </div>
   );
