+++ src/router/rootConstants.ts
@@ -10,6 +10,6 @@ export const settingsPath = '/settings';
 
 export const errorPath = 'error';
 
-export const creatingNewBooking = '/creatingBooking';
+export const creatingNewBookingPath = '/creatingBooking';
 
-export const signUpForm = '/signUp';
+export const signUpFormPath = '/signUp';
